+++ data/duke.txt
@@ -2,3 +2,6 @@
 [T][ ]bitches
 [D][ ]orgy  (by: Sep 8 2022 6pm)
 [E][ ]sex  (at: Aug 7 2021 5pm)
+[T][ ]gaysex
+[D][ ]gaysex  (by: Sep 8 2021 5pm)
+[E][ ]orgy  (at: Aug 7 2022 7pm)
