+++ src/main/java/DukeOperations.java
@@ -32,37 +32,7 @@ public class DukeOperations {
         System.out.println("_________________________________________________________________________");
     }
 
-    public void todo(ArrayList<Task> array, String[] arr, int counter) throws DukeException{
-        if (arr.length <= 1) {
-            throw new DukeException("☹ OOPS!!! The description of a todo cannot be empty.");
-        }
-        else {
-            array.add(counter, new Todo(arr[1]));
-            System.out.println("Got it. I've added this task:");
-            System.out.println(array.get(counter).toString());
 
-            counter++;
-            System.out.println("Now you have " + counter + " tasks in the list.");
-        }
-    }
-
-    public void deadline(ArrayList<Task> array,String item, String deadline, String timing,  int counter){
-
-        array.add(counter,new Deadline(item, deadline, timing));
-        System.out.println("Got it. I've added this task:");
-        System.out.println(array.get(counter).toString());
-        counter ++;
-        System.out.println("Now you have " + counter + " tasks in the list.");
-    }
-
-    public void event(ArrayList<Task> array,String item, String deadline, String timing, int counter) {
-        array.add(counter,new Event(item,deadline, timing));
-        System.out.println("Got it. I've added this task:");
-        System.out.println(array.get(counter).toString());
-        counter ++;
-        System.out.println("Now you have " + counter + " tasks in the list.");
-
-    }
 
     public void mark(ArrayList<Task> array, int number) {
         array.get(number-1).markAsDone();
@@ -85,15 +55,6 @@ public class DukeOperations {
 
     }
 
-    public void delete(ArrayList<Task> array, int number,int counter) {
 
-        System.out.println("Noted. I've removed this task:");
-        System.out.println(array.get(number-1).toString());
-        int num = counter - 1;
-        System.out.println("Now you have " + num + " tasks in the list");
-        array.remove(number - 1);
-
-
-    }
 
 }
